[space] SPACE-11367 Upgrade space-idea-sdk version to 1.1.58107
The Space Automation scripts (.space.kts) are compiled using a compiler downloaded from the Space server, and the evaluation of the script executes this compiled code by dynamically loading the compiled jar.
 The problem is that this compiled code uses the same classes as the embedded Space IDEA SDK, but in a different version.

Recent changes in the public API on the Space side made this discrepancy problematic. This is the root cause of the "when" statement failure, and actually still causes other issues.

IJ-MR-3820

GitOrigin-RevId: db174ab27642dd5323a3fd521459b3be34d412da