Making "in-place rename" not rely on the `RENAME_SEARCH_IN_COMMENTS_FOR_FILE` setting in tests
Because `RENAME_SEARCH_IN_COMMENTS_FOR_FILE` is not actually used in production code for the in-place rename. And having the `RENAME_SEARCH_IN_COMMENTS_FOR_FILE` set to `true` by default led to testing the "ALL_OCCURRENCES" mode by default, which is not that you usually want to test.

GitOrigin-RevId: 7147f761a838b4563e3d59598210b15b92d1ccc6