[platform] restore compatibility with code which use PlaceProvider and PlaceHolder (IDEA-254534)
Removal of generic parameter (in 09701d6d9) changed return type of 'getPlace' and parameter type of 'setPlace' from Object to String in bytecode so code compiled against IDEA 2020.3 won't be compatible with IDEA 2020.2 and older versions. This change restores methods with the old signatures in bytecode and delegates them to the new methods.

GitOrigin-RevId: b8d432d98c0a02f654db7558d0ca497eff720c37