Add support for parametrized model builders injected from the IDE
Gradle since version 4.4 supports ParameterizedToolingModelBuilder
interface which allows model builders to be configured form the IDE.
This is used by the Android plugin to import only the build variant
currently selected in the IDE.

This change make it possible to configure model builders injected from
the IDE in a similar manner. This should allow the Android plugin to
communicate the currently selected build variant to the Kotlin plugin
and thus not to import unnecessary source sets.

Change-Id: Ic04f52735eb56b300934602d81fda380f9660955