Rewrite get_wsl_ip to C and bundle binary.
It seems that every IJ product needs to know WSL-2 IP, but there are no guarantees for an interpreter/runtime to exist on WSL except glibc and kernel of course.

Binary tool runs everywhere.

It must be compiled statically on the oldest Linux possible. It gives us guarantee that it runs on distros without of glibc (i.e. Alpine).

GitOrigin-RevId: 92b26d4e176cc3a46ede7970a2b08a228d0f89b7