[properties] IDEA-215803 Join and split string works incorrectly inside property files when cursor inside property value
The enter handler for properties didn't used to take into account that there might be whitespaces where the user hit the enter key which resulted in losing these whitespaces in the property value. Now the enter handler checks if the caret at a whitespace symbol and if it is, then advances the caret to the last consequent whitespace.

The join lines handler for properties didn't take into account that if there is an indent in the next line which is not a part of the property value which resulted in altering the value with the indent whitespaces when a user joins the lines. Now the join lines handler considers all the leading whitespaces from the second line as its indent and didn't add its whitespaces as a meaningful symbols of a property's value

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 3f8fb9c1f2141152e9b14c47a31652673e73ef43