[properties] IDEA-229124 Properties don't align in column when key-value delimiter is space
IDEA didn't use to consider that during formatting the key and the value of a property can be separated by a whitespace and as a result didn't register a proper alignment when a whitespace is set as a separator in the code style settings. This patch adds a check if the lexer didn't find a key-value separator (either "=" or ":") and if so it registers an alignment on the property value.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 1126807dfd92fdab81e65ad8a60c947f1e914605