[vcs-log] create VcsLogUi in DefaultVcsLogFile
In this commit DefaultVcsLogFile keeps the id of the VcsLogUi to create instead of VcsLogUi instance. Exception class CannotAddVcsLogWindowException is introduced in case VcsLogUi by the same id was created earlier (since it is not prohibited to open the same file twice).

In the future this will allow to reopen the file from "Recent Files" popup.

GitOrigin-RevId: 0725e5cf6f3fcecbbff6fd3da217d7fac917e7b1