[vcs-log] introduce file system for vcs log editor tabs
This allows to reopen closed log tabs from the "Recent Files" popup.

After this change, on project start log tabs are going to be reopened by the platform, not by VcsLogTabsManager. But VcsLogTabsManager is also going to try to reopen the tabs, as it is necessary to do so on every VcsLogManager creation, which can happen not on start. In order to avoid opening the same files twice, VcsLogTabsManager now passes "searchForOpen" flag to the "openFile" method.

Another detail is that when log tabs are reopened on start, it is impossible to get VcsLogUi instance for them immediately. DefaultVcsLogFile waits for log initialization using "runWhenLogIsReady" method, and at the time when VcsLogTabsManager tries to recreate previously opened tabs, the callback from "runWhenLogIsReady" has not been executed yet. So "openLogTab" returning VcsLogUi instance could not be used as is here and had to be inlined.

GitOrigin-RevId: 44199c3f04a68a15930bea232ab5fd36ef5a603f