phm: split PersistentHashMap from it's implementation
yet another binary/source compatibility change. It moves the PersistentHashMap implementation
to a dedicated hierarchy of PersistentHashMapBase, introduces PersistentHashMapBuilder to
get rid of too many overloaded constructors.

The PersistentHashMap itself is now turned into a delegated implementation what
delegates to underlying PersistentHashMapBase instance. Most of the methods of
PersistentHashMap are now final to avoid a sudden replace

Some custom tunings of PersistentHashMap can now be broken (is seems there are no such cases
in external plugins, thanks to final methods, these are likely to fail fast), the best is to
delegate and wrap PersistentHashMap in the code. When impossible, extend PersistentHashMapImpl
and wrap the outcome with PersistentHashMap

GitOrigin-RevId: 3730c5a3899d0a542e60c416c8ce0c28e85b3444