CPP-12394: Fix INCLUDE_GUARD for file templates
INCLUDE_GUARD was not handled for custom file templates at all and was presented to the user in the dialog.
We want to generate the correct INCLUDE_GUARD automatically, instead of showing it to the user.
We follow the same rules for header guard generation as we have for "New C/C++ Header File".

GitOrigin-RevId: 13f8cd5679933722ebb3dd95f347087d825e90bd