[workspace model] make instances of OrderEntry survive after changes in ModifiableRootModelBridgeImpl (IDEA-255223, EA-237923)
It may happen that a client stores an instance returned from ModifiableRootModel::getOrderEntries and call some setters on it after changing list of order entries. We have such code at least in "Copy to module libraries" action and in Dependencies table in Project Structure dialog, there may be more cases in plugins. Before this commit such code may fail because OrderEntryBridge will use the original index of the entry to modify data in ModuleEntity.

 In order to fix this we need to keep instances returned from 'getOrderEntries' and update indices in them when list of order entries changes, it's done via ModifiableRootModelBridgeImpl::mutableOrderEntries. RootModelBridgeImpl snapshot isn't used to retrieve information related to order entries, it's still used for content and source roots though.

 This change also removes unnecessary properties from OrderEntriesBridge's implementation which store information which may be taken from 'item' property instead.

GitOrigin-RevId: deaeb7d1da42f7d398877885b8cbc250b2764af6