[properties] IDEA-215803 Join and split string works incorrectly inside property files when cursor inside property value
This patch amends PropertiesHighlighter with not highlighting invalid string escapes since Properties are meant to ignore backslashes if they don't precede a single quote, double quote or backslash character.

This patch amends EnterInPropertiesFileHandler with splitting a property value right at the place of the caret and escaping optional whitespaces with a backslash on the next line if the caret is at a whitespace.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 0d8b66c63db1d81a6deada3c26b808fa3b61d46b