[Windows COM] Use single thread apartment for EDT.
``TaskBarList`` supports ``Both`` threading model, so it can run on STA when accessed strongly from the same thread which was used to create it (hence, apratement is the same).
Since it is accessed from EDT only anyway, we can use STA.

Initializing COM in MTA mode breaks CEF offscreen mode because it depends on STA (it uses ``OleInitialize``).

Context is also simplified because only inproc server is used

GitOrigin-RevId: 11966437c7f60b0e0b1a148f69f0f32e3436b568