IDEA-255722: use the run configuration location to check if it's from a given context.
for "all in package" run configurations, we cannot rely on the name of
the package only, because this doesn't garantee that two configurations
belongs to the same context. For example: "instrumented tests" and "unit
tests" in Android Studio.

The fix adds the location information as userData when creating the run
configuration. Another solution would be to extend
ExternalSystemTaskExecutionSettings to store this information as well,
but I am not sure if this kind of extensions is easily approvable.