[workspace model] Add check store consistency before flushing on disk
It was decided to add a double-check of the store consistency before flushing to the disk.
This is due to the extraction of the check consistency process to the separate thread.
And there are some cases then addDiff or replaceBySource was already executed but the store
consistency was not calculated yet. Thus during store serialization, we will not get if the
store was actually corrupted.

GitOrigin-RevId: cde63e25d945539befbec392889547b843713db4