phm: default implementation for PersistentHashMapBase#appendData method
In the long run, it seems better to move the appendData method
deep into the PersistentHashMapImpl implementation details,
right now we need a stub implementation to go on with other
experiments. It is better to implement the method more
efficiently in any PersistentHashMapBase implementations.

It may turn out later we come up with yet another appendData
method that would take Value or Collection<Value> instead
of the abstraction-breaking ValueDataAppender

GitOrigin-RevId: 7df496cfaa859fe431af26559a26afafac25c46e