[properties] IDEA-248720 Sort properties files loses comments
This patch fixes the regular expression that checks if a whitespace token one or many consequent empty lines. The regular expression used to invoke System#lineSeparator to get the platform's newline symbol, but it turned out that PsiElements already have all the platform dependent newline symbols converted to a single '\n'.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 8fa908b4f43bde5a02644eeffb77b619f2d566dc