PY-45541 Display symbol's canonical import path as tail text in the lookup
Previously, we used the corresponding element's presentation for this purpose,
but it showed the shortest qualified name of its module, not necessarily the
"canonical" name that was actually used for importing. For instance, we could
have suggested `ndarray` from `numpy.core._multiarray_umath` but imported it
as `from numpy import ndarray`. It was both misleading and could have led to
duplicate entries for the same symbol among suggestions.

GitOrigin-RevId: abb350feeac9fd212d47dc0943442e0b16b21c9a