[workspace model] provide a way to save configuration of facets for modules imported from external sources
It may happen that a module configuration is imported from some external model, but user may manually configure a facet in such module (e.g. Python facet in CLion, see PY-45374). In that case we need to produce an *.iml file for such module to store the facet configuration. CustomModuleEntitySource is introduced to mark entity sources of such modules.

GitOrigin-RevId: db079434df6bf2684eaa5a82b0e0d49e41993806