[vcs] correctly identify toolwindow by tab name before toolwindow content initialization
After the introduction of the new api for specifying whether to use "Commit" toolwindow for the tab, it is impossible to identify the correct toolwindow just by the tab name. ChangesViewContentManager keeps track of tabs that should be placed in the "Commit" toolwindow, but only after "Commit" toolwindow content was initialized. Before that moment toolwindow for tabs were identified incorrectly, which broke navigation actions.

 This commit adds looking through ChangesViewContentEP implementations to identify the correct toolwindow before initialization.

 follow-up: c3aa1441d627dbf535a56314322d27abc3d37e3c

GitOrigin-RevId: 0ca86e9e842737f5ed9971fac17558cb50e872df