IDEA-256068 clear gist ID on plugin unloading
GistManagerImpl stores IDs of gists to track situations when the same ID is used for different gists. But it doesn't clear IDs on a dynamic plugin unloading. Now, in addition to an ID, a weak reference to a gist itself is stored. It's expected that when the plugin is unloaded, the gist is unloaded as well clearing the ID. Thus, when the plugin is loaded back, it can create the gist again using the same ID.

GitOrigin-RevId: beab4a10c27b2ef8daffee1a6c4834ec871c156a