[workspace model] introduce DummyParentEntitySource to allow skipping dummy entities in replaceBySource
If *.iml file contains facet configuration only, and the module dependencies configuration is obtained from some other source, we still need to create some ModuleEntity in JpsProjectSerializers because FacetEntity requires it, but we need to ensure that such dummy instance won't replace real ModuleEntity instance when replaceBySource will be called. So DummyParentEntitySource is introduced to mark entitySource of such entities. It's needed to support loading facet configurations in CLion (PY-45374), later we can use the same approach for modules imported from Gradle and other external systems.

GitOrigin-RevId: bc40c6119d1f8d4d7d512fdaef39e9342ce7dd41