vcs: reduce log level to warn
This method should return default value if the specified file not exist, but LOG.error throw assertion in test mode and fail the whole test. The callers of this method should handle case of non-existent file on their own (e.g. file may be deleted by another process or thread before the delayed call to this method happened).

GitOrigin-RevId: 1b1db260d1b05e8ec23a73d02790b0c9f2cccd45