[properties color scheme] IDEA-256474 Properties Color Scheme settings miss Valid string escape option
As a part of refactoring that was done in IDEA-215803 com.intellij.lang.properties.PropertiesHighlighter.PROPERTIES_VALID_STRING_ESCAPE was removed since it didn't have obvious connections. But the attribute was used in the configuration of properties color scheme. This patch restores PROPERTIES_VALID_STRING_ESCAPE and also fixes the lexer for properties file to be able to ignore the problems with invalid escape character sequence if it's not a unicode code point, because the java.util.Properties#load method does not treat a backslash character, \, before a non-valid escape character as an error.

This patch also fixes the i18n warnings in PropertiesHighlighter with using messagePointer instead of the message method so no string constants are stored in static members.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 1d91ed63dbbb26cf62afc0adf6717e54b489d9ec