PY-45835: Support execution on background thread.
Since PythonRunner is AsyncProgramRunner, any PythonCommandLineState must be ready to be called on any thread.

But History is not PythonCommandLineState: it is a platfrom class instead, so PythonRunner can only assume than PythonCommandLineState is ready to be called on any thread. Any other line state must be called on EDT.

GitOrigin-RevId: ed021d554d6b43b5d71ec71a5152ee60ff16e96b