[AndroidStudio] Add customised deprecation message for Android JUnit.
As we are moving to Gradle test runner for unit tests, we want to help
users understand the deprecation of Android JUnit through the UI.

the deprecated run configs will now be categorised as
UnknownRunConfiguration (final class) that we cannot extend to
customise. When clicking on the run button for such configs, an
exception is raised and displayed on a window, and we want to customise
the exception for Android JUnit run configurations.

bug: 165193589
Test: N/A.
Change-Id: I143862898a9febaac7e0ed7e726403a931f100d7