settings: make Export Settings action identify more correctly if the storage is a directory: IDEA-252588
The AppCodeStyleSettingsManager stores code style settings in the `code.style.schemes` XML file without the ".xml" extension. It is against convention, but it was made so by mistake, and we can't change it without proper migration.

OTOH Storage.value is a file iff a custom StateSplitter is not defined for the component, so the fix is to rely on this statement rather than to look into the xml extension.

This was broken in 094bac1e.

(cherry picked from commit dd642e9745376657e23f82b889bf0b943ce9fddb)

IJ-CR-4542

GitOrigin-RevId: 6748a775fb12f7540bde1badcf0173e6d2f11c8a