Scopes Settings: Align file color after it is added to scope
Initially, when file is added to the scope the color was updated only in the tree,
but tab color stayed unchanged. Now it is updated as soon as settings are saved.
Tree update is also aligned with that, now colors are updated in the tree and
in tabs section simultaneously

GitOrigin-RevId: 98937f7257e4db7ffc00d7ccbdcce1bbd6f957bc