Suppress debug logging for assumption violations
In the light test we are wrapping logging around test run only, but not setUp or teardown. Therefore when debug output is empty, we are logging nothing.

In heavy test for better diagnostics we are wrapping tests with setUp and tearDown and setUp is always debugging something, so buffer is not empty.

With current solution we just wont log anything if assumption violation happened.

GitOrigin-RevId: dc412ea9385a148a88bc465caa5b0547accc3a71