[python] fix removal of Python facets (PY-45949)
In workspace model facets and libraries are store in common storage, so currently if libraries are changes inside FacetManagerListener events which are fired synchronously inside WorkspaceModel::updateProjectModel, it leads to recursive invocation of updateProjectModel methods which doesn't work properly (changes made in the recursive call are lost and state of the model becomes inconsistent leading to exceptions). As a quick fix updating libraries is performed later.

GitOrigin-RevId: d90a612a1d2e80f6eb26a32b47fbacad917024ff