RUBY-18675 add ability to use JLabel with ComboBoxCompositeEditor
It can be useful to use a composite editor to add extra icons/actions (such as "file browse"), without also making the actual combo box editable.

So far the only way I have found to achieve this is still to use a custom editor for a combo box, as it's the only way to provide a custom renderer for the selected (non drop-down) item in the combo box. (When a combo box is not "editable" then it uses the same renderer for the selected item as with the drop down items.)

IJ-CR-19452

(cherry picked from commit 15d3df1a59a7cdf76e0872875795fb67922f7824)

GitOrigin-RevId: 5e32d9095b02048391d663268f3b23bf94191200