[build scripts] don't put same modules related to javac2 tasks to both Java plugin and platform (IDEA-257865)
When the same class is added to platform and a plugin, it may be loaded by different classloaders and it may lead to exceptions at runtime. There is no need to have intellij.java.guiForms.rt module in Java plugin since it's included to the platform anyway. And javac2.jar was kept in IDEA_HOME/lib for compatibility with Ant scripts produced by "Generate Ant Build" action. The new version (203.3) of "Ant Build Generation" plugin takes javac2.jar from plugins/java/lib so the JAR can be removed from the old location.

GitOrigin-RevId: e3c874c53becee1f8c5858f2eebcb206eab8aaca