Customize deprecation message for AndroidJUnit
As we are moving to Gradle test runner for unit tests, we want to help
users understand the deprecation of Android JUnit through the UI.

The deprecated run configs will now be categorised as
UnknownRunConfiguration (final class) that we cannot extend to
customise. When clicking on the run button for such configs, an
exception is raised and displayed on a window, and we want to customise
the exception for Android JUnit run configurations.

This re-lands change I143862898a9febaac7e0ed7e726403a931f100d7, that
was overwritten by the 203.3645.34 merge.

Bug: 165193589
Test: N/A.
Change-Id: Ibfb83935083792fff49372ffc8012557aae1d8a7