PY-46361 Introduce PySearchUtilBase.defaultSuggestionScope()
It additionally excludes python-skeletons from PySearchUtilBase.excludeSdkTestsScope()
result. Now it's used only in "extended" completion and auto-import quickfix, both
suggesting not yet imported symbols, modules and packages to a user, which already
filtered their results like that. Probably it makes sense to migrate similar places
such as "Go to ..." action providers to this scope as well.

GitOrigin-RevId: 7544fa4fbd3043995630fd85e37c5467256e74fd