PY-46381 Don't suggest internal parts of third-party packages to user
Namely, their bundled dependencies and tests.

Common names of test and "vendored" roots were collected based on
a base Anaconda interpreter.

Filtration is done over a module/package qualified name in its closest root,
sharing some parts of the implementation with QualifiedNameFinder.
It's likely to also be utilized in the upcoming auto-importing completion
of qualified names.

To make the customization of a search scope easier to use and extend in
the future, I introduced PySearchScopeBuilder API that allows to build
a custom search scope, excluding some irrelevant parts of a Python SDK.

I also updated the set of known standard library tests.
"idlelib/testcode.py" was removed as the only file entry that is
found only in Python 3.3, which we no longer support.

GitOrigin-RevId: 6676c59011d51371639ce24a5ac5c5b56d6b13fb