Fix typo in ProjectRootsChangeListener
Typo: "==" instead of "=". This bug causes that sporadically after adding roots they are not indexed.

If we have a sequence of events there is "remove" event  before "add" even, due to a bug a final type will be `RootsChangeType.ROOTS_REMOVED` even if we have "add" event. Therefor we never trigger an indexing for these changes.