Fix a typo in ProjectRootsChangeListener.kt
Typo: "==" instead of "=". This bug causes that sporadically after adding roots they are not indexed.

If we have a sequence of events there is a "remove" event before an "add" even, due to the typo a final type will be RootsChangeType.ROOTS_REMOVED even if we have an "add" event. Therefore we never trigger an indexing for these changes.

PR to JB: https://github.com/JetBrains/intellij-community/pull/1493

Fixes: 177070442
Test: tested on ComposeSampleResolutionServiceTest#testResolveSampleReference
Change-Id: I2b3b4ba444b5310f4e9f8929af85205e5c4855d0