jdk-downloader: cleanup the WSL support and items filtering
The JdkPredicate is now used (again) for both WSL and M1 package
of JDK. It's up to the callee to decide on the predicate.
There is a cache for Jdk items per JdkPredicate now.

JdkDownloadDialog builds a single models for WSL case
and flips the prepared model back and forth depending on
the selected target path

GitOrigin-RevId: 249203a6236a3ff8834ec87467bac4336740208b