update values in wsl distribution descriptor
fixing the wrong msdi value for PENGWIN distribution, this error generate error to user when they try to connect to wsl PENGWIN distribution. This bug is present on several tools of intellij. 

Removing unsupported wsl distribution "WLINUX" . this distribution was renamed as PENGWIN by WhitewaterFoundry project, to avoid problems with different distributions with same msdi.  (this is the root cause of wrong msdi value for PENGWIN)

adding some missed distribution. to allow to users work with Intellij on theses wsl linux distributions.