Fix typo in ProjectRootsChangeListener
Typo: "==" instead of "=". This bug causes that sporadically after adding roots they are not indexed.

If we have a sequence of events there is "remove" event  before "add" even, due to a bug a final type will be `RootsChangeType.ROOTS_REMOVED` even if we have "add" event. Therefor we never trigger an indexing for these changes.

close #1493

GitOrigin-RevId: 1f8346692e3386634835569536b440179f736dc2