[groovy] IDEA-229704: Handle recursive calls in special way
There is a problem with cycles in the call graph: they are inevitably lead to recursion prevention. It is assumed that such cycles are rare enough (unless it is a sort of a visitor pattern), but usual recursion (which is a loop over one vertex in the call graph) is quite a common case. So all recursive calls now are resolved directly in virtual method, and that should improve overall recursion prevention tolerance and possibly lead to more precise results.

GitOrigin-RevId: f5cdbe22dc573e34a5824804046d5a50a976f1a7