[workspace model] implement workaround to fix renaming of a module with custom SDK (IDEA-260145)
There is no need to substitute instance of Sdk if workspace model is used, because it refers to Sdks by name anyway. There is more generic problem with multiple diffs (see IDEA-260248), it'll be addressed later. This commit provides a workaround to get rid of the problem in the most common case.

GitOrigin-RevId: 68f109c1538ab0303cc2a8b0e0b75f4439971269