IDEA-260227 remove `GotoLambdaParameterHandler`
This handler was needed to show the doc as if the parameter was referenced. In the new implementation we show the same
doc info regardless whether the target under caret is referenced or declared, but this handler made GTDU believe
that the parameter was referenced, and GTDU was choosing navigation to the referenced target instead of showing usages.

GitOrigin-RevId: d1a4fab90c47fad5a13409faf5d80a686b245c63