[workspace storage] don't overwrite external mappings by values from unrelated removed entities
If a parameter in 'addDiff' call contains 'addEntity' and 'removeEntity' change for a entity with some external mapping, it shouldn't clear external mapping for existing unrelated entity from the original builder which uses the same EntityId. This is needed to properly implement merging of multiple project-level libraries with the same name (see IDEA-249398).

GitOrigin-RevId: d6ecc5edf73e9859116f1f209c512f8d2f9955c6