Make ModuleRedeclarator tests extend GeneratorTestCase to fix flaky tests
Functional generator3 tests rely on the environment variable GENERATOR3_TEST_MODE
being set. Normally, it's set automatically in GeneratorTestCase.setUpClass and
then propagated to children processes. Moreover, on the first call the result
of the corresponding check is cached. ModuleRedeclarator tests also use this check
by instantiating ModuleRedeclarator, but don't set the variable beforehand, so
it becomes incorrectly cached as false for subsequent tests of the skeleton generator.
Test were flaky because the order of tests varies from run to run.

GitOrigin-RevId: bbe2a3cbac43e821cd13e09c9bae702501ce4c81