[vcs-menu] Move Resolve conflicts action to the first level of the menu.
Conflicts can occur not only because of Rebase or Merge, but also because of Cherry-Pick, Revert and Unstash. It appears having a single action in the same location is more handy than an action in submenu of each action. Moreover, there is not Cherry-pick and Revert actions in hte main menu.

IDEA-257569

GitOrigin-RevId: bc3d383ab82fe44cb810cf3fc791508040aee440