IDEA-261098 `GotoVarTypeHandler`: don't target `var` variable identifier to the variable
This was needed to show the doc as if the variable was referenced. In the new implementation we show the same
doc info regardless whether the target under caret is referenced or declared, but this handler made GTDU believe
that the variable was referenced, and GTDU was choosing navigation to the referenced target instead of showing usages.

GitOrigin-RevId: 78e3554df87d564aeea04f707a0fa1309031912f