[new api usage] IDEA-257412 - Update API list for Java 16
This patch adds the new api16.txt file that contains all the classes and class members that were added in JDK 16 (that contain the since tag).

In order to make it easier to add new tests for the cases that detect if API from a later language level comparing to the current one is used the BaseApiUseTestCase class was added. Its job is to configure the mock JDK. The class describes its usage including adding new classes to the mock JDK in its javadoc.

GitOrigin-RevId: 101cef98c088ebe7086b87dc93affe85e41ecf4f