kotlin-ide migration: Normalization commit
This commit is consequence of not accurate migration to kotlin-ide repo.
These changes appeared as a result of split of big chunk of first 
commits in kotlin-ide repo. I don't know where from changes in this 
commit come from.

It could be changes of "bunch switch" commit. Or changes of other 
accidentally squashed commits

GitOrigin-RevId: 6d1e18c36e02a4ad228fe7ddc5bd056e98e1be36