Fetch and process models in separate phases
This change splits `fetchProjectBuildModels` into two phases:
(1) getting models via Gradle API
(2) preprocessing models and adding them to myAllModels

This change enables a later change to replace the loop that requests
models sequentially with a collection of Gradle nested build actions
which, in Gradle 6.8 and later,  can run in parallel.

This change may slightly increase the memory usage of the Gradle process
because we do not release references to the original model instances
when they are preprocessed by a custom serializer. However, such models
are usually not too large.

Change-Id: I36f8b8ecbfa3f138af680dc6fb23094755924cef