Extract fetchProjectBuildModels method
This change replaces two similar loops with a method call which
requests project level models for a given `GradleBuild` and adds
them to myAllModels collection.

A later changes in this pull request converts this sequential loop
into a collections of Gradle nested build actions which, in Gradle 6.8
and later, are executed in parallel.

Change-Id: Ic99003f6dddf14189f9d55f448d654ae32c53204

GitOrigin-RevId: d824893dd053bc675f896a7b071a98f21537991b