Request project models via parallel-processing capable API
This change replaces the loop that requests project models sequentially
with an invocation of `org.gradle.tooling.BuildController#run` which
tries to execute actions in parallel when connected to Gradle 6.8 or
later and runs them sequentially otherwise.

The received models are still serialized and collected on a single
thread in the same order.

Change-Id: I0a890936d972627afc0b7472c135aa911937a6b3