[preview] IDEA-259969 Reflective preview APIs are incorrectly reported as errors
This patch extracts PreviewFeatureVisitorBase to the top level class. In the visitor there used to be a special case for the new expression in order to reduce the number of warnings/errors, but it appeared to be useless, since the logic to deduce the preview feature element is too complicated. We might need to circle back to it if users complain about the number of warnings in their code.

GitOrigin-RevId: 3355caedc7eae3de197f652cc6af7ae605f64971