[git-index] do not undo staging operations from the original document editor
Staging operations are performed on the staged document built from the GitIndexVirtualFile. When such operation is performed, original document is typically opened in the editor. Because of this, UndoManagerImpl#commandStarted adds reference to the original document to current actions list and the command gets into undo stack of the original document. So when user tries to undo a change in the original document, staging operation is undone instead.

In order to fix this behavior, a new api method in the CommandProcessor is introduced to pass the "shouldRecordCommandForActiveDocument" flag along with the modified document and skip reference to the original document addition. In order to pass this flag only for staged documents, SEPARATE_UNDO_STACK key is used to mark such documents.

This behavior is controlled by the "git.stage.separate.undo.stack" registry key.

IDEA-262282

GitOrigin-RevId: b96f19b5251c425252f2fb472bd4cb48988ea599