[workspace model] fix equals for ModuleImlFileEntitiesSerializer (IDEA-259705)
The previous variant returned 'true' when ModuleImlFileEntitiesSerializer is compared with ExternalModuleImlFileEntitiesSerializer created for the same module, this may lead to incorrect data in JpsProjectSerializersImpl::moduleSerializers, as a result module lists weren't updated properly.

We now save modules.xml in the both internal and external locations if any of these files should be saved to properly handle cases when module is moved from one of these locations to another.

GitOrigin-RevId: 1440004799cd5650c3efbc481ec2158f6a9d32f8