Treat project as trusted if it has settings in the APP_CONFIG/workspace: IDEA-262365
If the project has settings in APP_CONFIG/workspace, but its trusted state is unknown, that means that this project was already used on this computer before and it is safe to assume that it is trusted. This helps to distinguish old projects from new projects which have .idea, and to avoid showing the untrusted code confirmations for old projects.

GitOrigin-RevId: 2397ae0c3ab4334e20d974c4b3e58347f5ff8ce7