Treat project as trusted if it has settings in the APP_CONFIG/workspace: IDEA-262365
If the project has settings in APP_CONFIG/workspace, but its trusted state is unknown, that means that this project was already used on this computer before and it is safe to assume that it is trusted. This helps to distinguish old projects from new projects which have .idea, and to avoid showing the untrusted code confirmations for old projects.

(cherry picked from commit 2397ae0c3ab4334e20d974c4b3e58347f5ff8ce7)

IJ-CR-6145

GitOrigin-RevId: fba7e3e400d6f9a20356b1411d4d29340a560cc0