Show the notification panel only for untrusted Maven projects
The notification was shown for all untrusted projects, but in many cases they don't need to be trusted to be fully operational (e.g. JPS projects or web projects), so the notification is useful and confusing.

The notification will be shown for Gradle projects as well in a future commit.

Relates to IDEA-262197

(cherry picked from commit c36f40eb5ec689a0d235dfb25690f4d34b332037)

IJ-CR-6134

GitOrigin-RevId: 9990f6446134cf6a62c00f72f43a4e6fd1a88c19