`textSearch.kt#buildTextQuery`: an occurrence should be plaintext if there are no comments/strings at the offset at all
Before the change an occurrence in comment was considered as plaintext because `isApplicableSearchContext` returned
`false` for comment element, and then it returned `true` for the parent element of the comment, and the whole context
check was `true`, despite the occurrence was inside a comment. After the change if there is a comment or string element
in the path from a leaf to a file at a given offset, then the occurrence is not considered as plaintext.

GitOrigin-RevId: f30cc0526946c35c047a7567bed631e8615e1fab