Limit applicability of KotlinScriptDependenciesClassFinder to script files
'PsiElementFinder's are global and can be called for any context. As 'KotlinScriptDependenciesClassFinder' is meant to provide additional dependencies only for scripts, we need to know if the caller came from a script resolution context. Unfortunately, all we have here is a 'GlobalSearchScope', so we check it. (Ab)using scopes such a way is obviously not ideal, and, in the future, we might want to migrate to a completely different script handling approach (for instance, treating each script as an individual module).

GitOrigin-RevId: b361ad280f4cf071e4b18c8744c27dc99c0c642c