[java] artifact editor: fix searching for MANIFEST.MF in Maven/Gradle projects (IDEA-99596)
Resource files located under source roots aren't copied to the output in Maven/Gradle project by default, so if user creates MANIFEST.MF under main/java directory (instead of main/resources) it won't be copied to the artifact. So artifact editor shouldn't show such file as the JAR manifest to avoid confusion.

The new implementation is not very reliable: it's possible to mark main/java as a resource directory explicitly, and it's possible to specify pattern for resource directory which will filter out MANIFEST.MF file. However it works in regular cases.

GitOrigin-RevId: 2fdf7a4c704641dcd64fb1e791595d04f93c6e5f