[file-history] postpone walking deleted parts of the history graph
This commit continues solving the problem with not detected renames (for example, in non-trivial merge commits, as described in d6643e6b9c90812cdbea7733d8a87ecc07cc6f1c). After refiner algorithm passes such commit, it marks a big part of the graph incorrectly as deleted, which leads to missing commits in file history. Commit d6643e6b9c90812cdbea7733d8a87ecc07cc6f1c partially solves the problem by changing order of processing commit parents. But it does not help in all cases, for example if the offending commit is not a merge.

This commit postpones walking into the nodes where file was deleted until other commits are processed. Because of this, we are guaranteed to process the subgraph where file exists first (and thus find all of the detected renames there).

After this change, ordering of commit parents implemented in d6643e6b9c90812cdbea7733d8a87ecc07cc6f1c is no longer required, but it seems like a good idea to leave it in place.

IDEA-245949

GitOrigin-RevId: 1f0da83920c8b112e311eff2856158caa8fd9039