KT-32444: Work around UAST class init deadlock
The deadlock occurs because UastBinaryExpressionWithTypeKind has
a static field referencing a subclass. Thus the class initialization
locks for the superclass and the subclass may be acquired in
different orders on different threads.

We can work around this by eagerly initializing
UastBinaryExpressionWithTypeKind before its subclasses are referenced.
This is not a clean fix, but the only alternative is to change the
API of UastBinaryExpressionWithTypeKind in IntelliJ.

GitOrigin-RevId: 98444676a6690fbcb2203f53af27bb9f1656888a